refactor: make create household standalone component
