+++ lib/create_household/bloc/create_household_state.dart
@@ -1,0 +1,22 @@
+part of 'create_household_bloc.dart';
+
+class CreateHouseholdState extends Equatable {
+  const CreateHouseholdState({
+    this.households = const [],
+  });
+
+  final List<Household> households;
+
+  CreateHouseholdState copyWith({
+    List<Household>? households,
+  }) {
+    return CreateHouseholdState(
+      households: households ?? this.households,
+    );
+  }
+
+  @override
+  List<Object?> get props => [
+        households,
+      ];
+}
