+++ lib/create_household/create_household.dart
@@ -1,0 +1,1 @@
+export 'view/view.dart';
