+++ lib/create_household/view/view.dart
@@ -1,0 +1,1 @@
+export 'create_household.dart';
